move the location update to the register pkg
