+++ pkg/registrar/registrar.go
@@ -13,19 +13,20 @@ import (
 	"github.com/rs/zerolog/log"
 	substrate "github.com/threefoldtech/tfchain/clients/tfchain-client-go"
 	"github.com/threefoldtech/zbus"
-	"github.com/threefoldtech/zos/pkg"
 	"github.com/threefoldtech/zos/pkg/app"
 	"github.com/threefoldtech/zos/pkg/environment"
+	"github.com/threefoldtech/zos/pkg/geoip"
 	"github.com/threefoldtech/zos/pkg/stubs"
 	"github.com/threefoldtech/zos4/pkg/geoip"
 )
 
 // should any of this be moved to pkg?
+type RegistrationState string
 
 const (
-	Failed     pkg.RegistrationState = "Failed"
-	InProgress pkg.RegistrationState = "InProgress"
-	Done       pkg.RegistrationState = "Done"
+	Failed     RegistrationState = "Failed"
+	InProgress RegistrationState = "InProgress"
+	Done       RegistrationState = "Done"
 
 	monitorAccountEvery    = 30 * time.Minute
 	updateLocationInterval = 24 * time.Hour
@@ -36,8 +37,15 @@ var (
 	ErrFailed     = errors.New("registration failed")
 )
 
-func FailedState(err error) pkg.State {
-	return pkg.State{
+type State struct {
+	NodeID uint32
+	TwinID uint32
+	State  RegistrationState
+	Msg    string
+}
+
+func FailedState(err error) State {
+	return State{
 		0,
 		0,
 		Failed,
@@ -45,8 +53,8 @@ func FailedState(err error) pkg.State {
 	}
 }
 
-func InProgressState() pkg.State {
-	return pkg.State{
+func InProgressState() State {
+	return State{
 		0,
 		0,
 		InProgress,
@@ -54,8 +62,8 @@ func InProgressState() pkg.State {
 	}
 }
 
-func DoneState(nodeID uint32, twinID uint32) pkg.State {
-	return pkg.State{
+func DoneState(nodeID uint32, twinID uint32) State {
+	return State{
 		nodeID,
 		twinID,
 		Done,
@@ -64,13 +72,13 @@ func DoneState(nodeID uint32, twinID uint32) pkg.State {
 }
 
 type Registrar struct {
-	state pkg.State
+	state State
 	mutex sync.RWMutex
 }
 
 func NewRegistrar(ctx context.Context, cl zbus.Client, env environment.Environment, info RegistrationInfo) *Registrar {
 	r := Registrar{
-		pkg.State{
+		state: State{
 			0,
 			0,
 			InProgress,
@@ -83,13 +91,13 @@ func NewRegistrar(ctx context.Context, cl zbus.Client, env environment.Environme
 	return &r
 }
 
-func (r *Registrar) setState(s pkg.State) {
+func (r *Registrar) setState(s State) {
 	r.mutex.Lock()
 	defer r.mutex.Unlock()
 	r.state = s
 }
 
-func (r *Registrar) GetState() pkg.State {
+func (r *Registrar) getState() State {
 	r.mutex.RLock()
 	defer r.mutex.RUnlock()
 	return r.state
@@ -205,11 +213,11 @@ func (r *Registrar) updateLocation(ctx context.Context, cl zbus.Client) error {
 }
 
 func (r *Registrar) NodeID() (uint32, error) {
-	return r.returnIfDone(r.GetState().NodeID)
+	return r.returnIfDone(r.getState().NodeID)
 }
 
 func (r *Registrar) TwinID() (uint32, error) {
-	return r.returnIfDone(r.GetState().TwinID)
+	return r.returnIfDone(r.getState().TwinID)
 }
 
 func (r *Registrar) returnIfDone(v uint32) (uint32, error) {
